LayoutTests/imported/w3c/web-platform-tests/fetch/sec-metadata/redirect/same-site-redirect.tentative.https.sub-expected.txt



FAIL Same-Site -> Same-Origin redirect assert_equals: destination expected (string) "image" but got (undefined) undefined
FAIL Same-Site -> Same-Site redirect assert_equals: destination expected (string) "image" but got (undefined) undefined
FAIL Same-Site -> Cross-Site redirect assert_equals: destination expected (string) "image" but got (undefined) undefined
